modified make file to have unit test command added .idea to gitignore fixed unit test files to not throw warnings on using gym.spaces

+++ all/approximation/action/discrete_linear_test.py
@@ -6,7 +6,7 @@ from all.approximation.bases.fourier import FourierBasis
 
 NUM_ACTIONS = 3
 LEARNING_RATE = 0.1
-SPACE = Box(low=0, high=1, shape=(2,))
+SPACE = Box(low=0, high=1, shape=(2,), dtype=np.float32)
 BASIS = FourierBasis(SPACE, 2)
 STATE = np.array([0.5, 1])
